refactor: move ZAP hook script to external template file
Separated the ZAP authentication hook script from hardcoded f-strings
in Python code to an external template file for better maintainability.

- Created resources/templates/zap_hooks_template.py with placeholder format strings
- Modified _create_auth_hook_script() to load and use the template
- Uses .format() method to replace placeholders with configuration values

This makes the hook script easier to maintain and customize without
modifying the scanner code.
